init schema editor component
