+++ app/lib/components/schema-editor/README.md
@@ -1,0 +1,1 @@
+# Visual Schema editor for defining custom schema for ui components
