+++ app/lib/components/schema-editor/index.ts
@@ -1,0 +1,1 @@
+export { SchemaEditor } from "./schema-editor";
