+++ app/lib/components/schema-editor/schema-editor.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function SchemaEditor() {
+  return <>schema editor</>;
+}
