+++ app/lib/screens/component-view/component-codebox.tsx
@@ -1,0 +1,46 @@
+import React from "react";
+import { CodeboxEditDialog } from "../../components/codebox-edit-dialog";
+import CodeBox from "../../components/codebox";
+import Settings from "@material-ui/icons/Settings";
+import IconButton from "@material-ui/core/IconButton/IconButton";
+
+export function ComponentCodebox(props: {
+  onCodeChange: (code: string) => void;
+  code: string;
+}) {
+  const [dopened, setdOpen] = React.useState(false);
+  const [code, setCode] = React.useState(props.code);
+
+  const handleClose = () => {
+    setdOpen(false);
+  };
+
+  const handleEditCodeClick = () => {
+    setdOpen(true);
+  };
+  return (
+    <div>
+      <CodeboxEditDialog
+        initialValue={code}
+        handleSave={(c) => {
+          setCode(c);
+          props.onCodeChange(c);
+        }}
+        label="code"
+        title="override code"
+        description="override the component code"
+        open={dopened}
+        handleClose={handleClose}
+      />
+      <CodeBox
+        codeActions={[
+          <IconButton onClick={handleEditCodeClick}>
+            <Settings />
+          </IconButton>,
+        ]}
+        language="tsx"
+        code={code ?? "//no code snippet provided"}
+      />
+    </div>
+  );
+}
