+++ app/lib/screens/component-view/schema-editor.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import { SchemaEditor } from "../../components/schema-editor";
+import { schema } from "coli";
+
+export function ComponentSchemaEditor() {
+  return <SchemaEditor />;
+}
